rename deepoctl -> deepocli
