+++ deepocli/cmds/blur.py
@@ -7,21 +7,21 @@ import logging
 import progressbar
 import numpy as np
 
-import deepoctl.cmds.infer as infer
-import deepoctl.io_data as io_data
-import deepoctl.workflow_abstraction as wa
+import deepocli.cmds.infer as infer
+import deepocli.io_data as io_data
+import deepocli.workflow_abstraction as wa
 
-class DrawThread(infer.InferenceThread):
+
+class BlurThread(infer.InferenceThread):
     def __init__(self, input_queue, output_queue, **kwargs):
-        super(DrawThread, self).__init__(input_queue, output_queue, **kwargs)
-        self.process = io_data.DrawOutputData(**kwargs)
+        super(BlurThread, self).__init__(input_queue, output_queue, **kwargs)
+        self.process = io_data.BlurOutputData(**kwargs)
 
     def processing(self, name, frame, prediction):
         return self.process(name, frame, prediction)
 
-
 def main(args, force=False):
     try:
-        io_data.input_loop(args, DrawThread)
+        io_data.input_loop(args, BlurThread)
     except KeyboardInterrupt:
         pass
