+++ deepocli/cmds/studio_helpers/image.py
