--- deepoctl/common.py
@@ -1,3 +1,0 @@
-
-class DeepoCTLException(Exception):
-    pass
